requirements-api.txt
openai
numpy
scipy
pytest
python-dotenv
sentence-transformers
matplotlib
pandas
fastapi
uvicorn
